\name{K.list}
\alias{K.list}
\docType{data}
\title{Example dataset, a list of kinship matrices}
\description{
a list object of the kinship matrices. It has 4 elements for 4 study cohorts. Each element is a kinship matrix. %~
}
\usage{data(K.list)}
\examples{
data(K.list)

length(K.list)

head(K.list[[1]])
}
\keyword{datasets}

