\name{y.list}
\alias{y.list}
\docType{data}
\title{Example dataset, a list of phenotype vectors}
\description{
a list object of continuous phenotypes. It has 4 elements for 4 study cohorts. Each element is a vector of continuous phenotypes %~
}
\usage{data(y.list)}
\examples{
data(y.list)

length(y.list)

length(y.list[[1]])
}

\keyword{datasets}

